자기 글 만 수정, 삭제 가능

+++ src/main/java/com/prj4/controller/board/BoardController.java
@@ -3,6 +3,7 @@ package com.prj4.controller.board;
 import com.prj4.domain.board.Board;
 import com.prj4.service.board.BoardService;
 import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.security.core.Authentication;
@@ -50,12 +51,25 @@ public class BoardController {
     }
 
     @DeleteMapping("{id}")
-    public void delete(@PathVariable Integer id) {
-        service.delete(id);
+    @PreAuthorize("isAuthenticated()")
+    public ResponseEntity delete(@PathVariable Integer id, Authentication authentication) {
+        if (service.hasAccess(id, authentication)) {
+            service.delete(id);
+            return ResponseEntity.ok().build();
+        }
+        return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
     }
 
+
     @PutMapping("edit")
-    public ResponseEntity edit(@RequestBody Board board) {
+    @PreAuthorize("isAuthenticated()")
+    public ResponseEntity edit(@RequestBody Board board,
+                               Authentication authentication) {
+        if (!service.hasAccess(board.getId(), authentication)) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+        }
+
+
         if (service.validate(board)) {
             service.edit(board);
             return ResponseEntity.ok().build();
@@ -65,6 +79,11 @@ public class BoardController {
 
 
     }
-
 }
+
+
+
+
+
+
